refactor(group): group shortcuts move to group-plugin
