+++ apps/demo-free-layout/src/components/selector-box-popover/index.tsx
@@ -5,6 +5,7 @@ import { Button, ButtonGroup, Tooltip } from '@douyinfe/semi-ui';
 import { IconCopy, IconDeleteStroked, IconExpand, IconShrink } from '@douyinfe/semi-icons';
 
 import { FlowCommandId } from '../../shortcuts/constants';
+import { WorkflowGroupCommand } from '../../plugins/free-group-plugin/constant';
 import { IconGroup } from '../../assets/icon-group';
 
 const BUTTON_HEIGHT = 24;
@@ -62,7 +63,7 @@ export const SelectorBoxPopover: FunctionComponent<SelectorBoxPopoverProps> = ({
             type="primary"
             theme="solid"
             onClick={() => {
-              commandRegistry.executeCommand(FlowCommandId.GROUP);
+              commandRegistry.executeCommand(WorkflowGroupCommand.Group);
             }}
           />
         </Tooltip>
