+++ apps/demo-free-layout/src/plugins/free-group-plugin/constant.ts
@@ -1,0 +1,4 @@
+export enum WorkflowGroupCommand {
+  Group = 'group',
+  Ungroup = 'ungroup',
+}
