+++ apps/demo-free-layout/src/plugins/free-group-plugin/shortcuts/index.ts
@@ -1,0 +1,2 @@
+export { GroupShortcut } from './group';
+export { UngroupShortcut } from './ungroup';
