+++ apps/demo-free-layout/src/plugins/free-group-plugin/shortcuts/ungroup.ts
@@ -6,11 +6,11 @@ import {
   WorkflowSelectService,
 } from '@flowgram.ai/free-layout-editor';
 
-import { FlowCommandId } from '../constants';
-import { WorkflowGroupService } from '../../plugins/free-group-plugin';
+import { WorkflowGroupService } from '../workflow-group-service';
+import { WorkflowGroupCommand } from '../constant';
 
 export class UngroupShortcut implements ShortcutsHandler {
-  public commandId = FlowCommandId.UNGROUP;
+  public commandId = WorkflowGroupCommand.Ungroup;
 
   public commandDetail: ShortcutsHandler['commandDetail'] = {
     label: 'Ungroup',
