+++ apps/demo-free-layout/src/shortcuts/shortcuts.ts
@@ -2,10 +2,8 @@ import { FreeLayoutPluginContext, ShortcutsRegistry } from '@flowgram.ai/free-la
 
 import { ZoomOutShortcut } from './zoom-out';
 import { ZoomInShortcut } from './zoom-in';
-import { UngroupShortcut } from './ungroup';
 import { SelectAllShortcut } from './select-all';
 import { PasteShortcut } from './paste';
-import { GroupShortcut } from './group';
 import { ExpandShortcut } from './expand';
 import { DeleteShortcut } from './delete';
 import { CopyShortcut } from './copy';
@@ -20,8 +18,6 @@ export function shortcuts(shortcutsRegistry: ShortcutsRegistry, ctx: FreeLayoutP
     new ExpandShortcut(ctx),
     new DeleteShortcut(ctx),
     new ZoomInShortcut(ctx),
-    new ZoomOutShortcut(ctx),
-    new GroupShortcut(ctx),
-    new UngroupShortcut(ctx)
+    new ZoomOutShortcut(ctx)
   );
 }
